Ensure that the <category> tags are unique in feed
If a post is in a category with the same name as one of the tags, there
will be duplicates in the `<category>` tags in the Atom feed. This
feature ensures that the categories are unique in a case-insensitive
manner.

NOTE: This is a breaking changes. **This will translate all
tags/categories to _lowercase_ in the Atom feed.**